VGF-162 make drt_advance_book_min a double

+++ src/main/java/org/opentripplanner/index/model/TripShort.java
@@ -17,7 +17,7 @@ public class TripShort {
     public Integer direction;
     public String drtMaxTravelTime;
     public String drtAvgTravelTime;
-    public Integer drtAdvanceBookMin;
+    public Double drtAdvanceBookMin;
 
     // INCLUDE start and end time, pattern and route in detail version
